% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcapcPeaks.r
\name{gcapcPeaks}
\alias{gcapcPeaks}
\title{GC Effects Aware Peak Calling}
\usage{
gcapcPeaks(cov, gcbias, bdwidth = 200L, flank = round(bdwidth/2),
  prefilter = 4L, permute = 10L, pv = 0.05, plot = FALSE,
  genome = "hg19")
}
\arguments{
\item{cov}{A list object returned by function \code{read5endCoverage}.}

\item{gcbias}{A list object returned by function \code{gcEffects}.}

\item{bdwidth}{A non-negative integer specifying ChIP-seq binding width.
Usually generated by function \code{bindingWidth}. A bad estimation of 
bdwidth results no meaning of downstream analysis.}

\item{flank}{A non-negative integer specifying the flanking width of
ChIP-seq binding. This parameter provides the flexibility that reads
appear in flankings by decreased probabilities as increased distance
from binding region. This paramter helps to define the weight of GC
content calculation. Default is the half size of \code{bindingWidth}. Note
that flank size won't largely affect results due to weighted strategy.}

\item{prefilter}{A non-negative integer specifying the minimum of reads
to qualify a potential binding region. Regions with total of reads from
forward and reverse strands larger or equivalent to \code{prefilter} are
selected for downstream analysis. Default is 4.}

\item{permute}{A non-negative integer specifying times of permutation to
be performed. Default is 10. When whole large genome is used, such as
human, 5 times of permutation could be enough.}

\item{pv}{A numeric specifying p-value cutoff for significant
binding peaks. Default is 0.05.}

\item{plot}{A logical vector which, when TRUE (default), returns density
plots of real and permutation enrichment scores.}

\item{genome}{A \link[BSgenome]{BSgenome} object containing the sequences
of the reference genome that was used to align the reads, or the name of
this reference genome specified in a way that is accepted by the
\code{\link[BSgenome]{getBSgenome}} function defined in the \pkg{BSgenome}
software package. In that case the corresponding BSgenome data package
needs to be already installed (see \code{?\link[BSgenome]{getBSgenome}} in
the \pkg{BSgenome} package for the details).}
}
\value{
A GRanges of peaks with meta columns:
\item{es}{Estimated enrichment score.}
\item{pv}{p-value.}
}
\description{
This function calls ChIP-seq peaks using potential GC effects information.
Enrichment scores are calculated on sliding windows of prefiltered
large regions, with GC effects considered. Permutation analysis is
used to determine significant binding peaks.
}
\examples{
bam <- system.file("extdata", "chipseq.bam", package="gcapc")
cov <- read5endCoverage(bam)
bdw <- bindingWidth(cov)
gcb <- gcEffects(cov, bdw, samp = 0.15)
gcapcPeaks(cov, gcb, bdw)
}

