% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcEffects.r
\name{gcEffects}
\alias{gcEffects}
\title{ChIP-seq GC Effects Estimation}
\usage{
gcEffects(cov, bdwidth = 200L, flank = round(bdwidth/2), samp = 0.05,
  gcrange = c(0.3, 0.8), plot = TRUE, mu0 = 1, mu1 = 50, p = 0.02,
  converge = 1e-05, emtrace = TRUE, species = "Hsapiens",
  build = "hg19")
}
\arguments{
\item{cov}{A list object returned by function \code{read5endCoverage}.}

\item{bdwidth}{A non-negative integer specifying ChIP-seq binding width.
Usually generated by function \code{bindingWidth}. A bad estimation of 
bdwidth results no meaning of downstream analysis.}

\item{flank}{A non-negative integer specifying the flanking width of
ChIP-seq binding. This parameter provides the flexibility that reads
appear in flankings by decreased probabilities as increased distance
from binding region. This paramter helps to define the weight of GC
content calculation. Default is the half size of \code{bindingWidth}. Note
that flank size won't largely affect results due to weighted strategy.}

\item{samp}{A numeric specifying the proportion of regions to be
randomized for GC effects estimation. Default is 0.05.}

\item{gcrange}{A non-nagative numeric vector with length 2. This vector
set the range of GC content to filter regions. For human, most regions
have GC content between 0.3 and 0.8, which is set as the default. Other
regions with GC content beyond this range will be ignored.}

\item{plot}{A logical vector which, when TRUE (default), returns plots of
intermediate results.}

\item{mu0}{A non-negative numeric initiating read count signals for
background regions. This is treated as a starting value for EM algorithm.
Default is 1.}

\item{mu1}{A non-negative numeric initiating read count signals for
foreground regions. This is treated as a starting value for EM algorithm.
Default is 50.}

\item{p}{A non-negative numeric specifying the proportion of foreground
regions in all estimated regions. This is treated as a starting value for
EM algorithm. Default is 0.02.}

\item{converge}{A non-negative numeric specifying the condition of EM
algorithm termination. EM algorithm stops when the ratio of log likelihood
increment to whole log likelihood is less or equivalent to 
\code{converge}.}

\item{emtrace}{A logical vector which, when TRUE (default), allows to 
print the trace of log likelihood changes in EM iterations.}

\item{species}{A character specifying the symbol of species on which
ChIP-seq is generated. It will be used to extract DNA sequence
from BSgenome. For example, "Hsapiens" for human and "Mmusculus" 
for mouse.}

\item{build}{A character specifying the ucsc genome build of given 
species, such as "hg19" for "Hsapiens" and "mm10" for "Mmusculus".}
}
\value{
A list of objects
\item{glm0}{Estimated generalized linear model for background GC effects.}
\item{glm1}{Estimated generalized linear model for foreground GC effects.}
\item{mu0}{Predicted signals by GC content if belongs to background.}
\item{mu1}{Predicted signals by GC content if belongs to foreground.}
\item{z}{Estimated probabilty of being foreground for input windows.}
\item{p}{Estimated proportion of foreground regions.}
\item{ll}{Converged log likelihood.}
}
\description{
GC effects are estimated based on weighted GC content and reads count
on genome-wide windows, using generalized linear mixture models. Genome
wide windows are randomly sampled with given proportions. GC effects of
background and foreground are estimated separately, while effects on
background level act more likely to be the real bias.
}
\examples{
bam <- system.file("extdata/chipseq.bam",package="gcapc")
cov <- read5endCoverage(bam)
bdw <- bindingWidth(cov)
gcb <- gcEffects(cov,bdw,samp = 0.15)
}

