% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcEffects.r
\name{gcEffects}
\alias{gcEffects}
\title{ChIP-seq GC Effects Estimation}
\usage{
gcEffects(cov, bdwidth, flank = NULL, samp = 0.05, plot = TRUE,
  gcrange = c(0.3, 0.8), mu0 = 1, mu1 = 50, p = 0.02,
  converge = 0.001, emtrace = TRUE, genome = "hg19")
}
\arguments{
\item{cov}{A list object returned by function \code{read5endCoverage}.}

\item{bdwidth}{A non-negative integer vector with two elements 
specifying ChIP-seq binding width and peak detection half window size.
Usually generated by function \code{bindWidth}. A bad estimation of
bdwidth results no meaning of downstream analysis.}

\item{flank}{A non-negative integer specifying the flanking width of
ChIP-seq binding. This parameter provides the flexibility that reads
appear in flankings by decreased probabilities as increased distance
from binding region. This paramter helps to define effective GC
content calculation. Default is NULL, which means this paramater will
be calculated from \code{bdwidth}. However, if customized numbers 
provided, there won't be recalucation for this parameter; instead, the
2nd elements of \code{bdwidth} will be recalculated based on \code{flank}.}

\item{samp}{A numeric specifying the proportion of regions to be
randomized for GC effects estimation. Default is 0.05.}

\item{plot}{A logical vector which, when TRUE (default), returns plots
of intermediate results.}

\item{gcrange}{A non-nagative numeric vector with length 2. This vector
set the range of GC content to filter regions. For human, most regions
have GC content between 0.3 and 0.8, which is set as the default. Other
regions with GC content beyond this range will be ignored.}

\item{mu0}{A non-negative numeric initiating read count signals for
background regions. This is treated as a starting value for EM algorithm.
Default is 1.}

\item{mu1}{A non-negative numeric initiating read count signals for
foreground regions. This is treated as a starting value for EM algorithm.
Default is 50.}

\item{p}{A non-negative numeric specifying the proportion of foreground
regions in all estimated regions. This is treated as a starting value for
EM algorithm. Default is 0.02.}

\item{converge}{A non-negative numeric specifying the condition of EM
algorithm termination. EM algorithm stops when the ratio of log likelihood
increment to whole log likelihood is less or equivalent to 
\code{converge}.}

\item{emtrace}{A logical vector which, when TRUE (default), allows to 
print the trace of log likelihood changes in EM iterations.}

\item{genome}{A \link[BSgenome]{BSgenome} object containing the sequences
of the reference genome that was used to align the reads, or the name of
this reference genome specified in a way that is accepted by the
\code{\link[BSgenome]{getBSgenome}} function defined in the \pkg{BSgenome}
software package. In that case the corresponding BSgenome data package
needs to be already installed (see \code{?\link[BSgenome]{getBSgenome}} in
the \pkg{BSgenome} package for the details).}
}
\value{
A list of objects
\item{glm0}{Estimated generalized linear model for background GC effects.}
\item{glm1}{Estimated generalized linear model for foreground GC effects.}
\item{mu0}{Predicted signals by GC content if belongs to background.}
\item{mu1}{Predicted signals by GC content if belongs to foreground.}
\item{z}{Estimated probabilty of being foreground for input windows.}
\item{p}{Estimated proportion of foreground regions.}
\item{ll}{Converged log likelihood.}
}
\description{
GC effects are estimated based on effective GC content and reads count
on genome-wide windows, using generalized linear mixture models. Genome
wide windows are randomly sampled with given proportions. GC effects of
background and foreground are estimated separately, while effects on
background level act more likely to be the real bias.
}
\examples{
bam <- system.file("extdata", "chipseq.bam", package="gcapc")
cov <- read5endCoverage(bam)
bdw <- bindWidth(cov)
gcb <- gcEffects(cov, bdw, samp = 0.15)
}

